feat: enhance socket authentication and update project file operations to include userId

+++ src/sockets/projectSink.ts
@@ -1,12 +1,12 @@
+import mongoose from "mongoose";
 import { ProjectFile } from "../model/project_files.js";
 import { Socket } from "socket.io";
 
-const projectUpdate = async (path: string, fsContent: string, name: string): Promise<boolean> => {
+const projectUpdate = async (path: string, fsContent: string, name: string, userId: mongoose.Types.ObjectId): Promise<boolean> => {
   try {
     const projectFile = await ProjectFile.findOneAndUpdate(
-      { path, name },
+      { path, name, userId },
       { content: fsContent },
-      { upsert: true, new: true }
     );
 
     return !!projectFile;
@@ -20,7 +20,8 @@ export const projectSocket = (socket: Socket) => {
   // Update file content
   socket.on("file:update", async (req) => {
     const { path, fsContent, name } = req;
-    const success = await projectUpdate(path, fsContent, name);
+    const userId = socket.data.userId;
+    const success = await projectUpdate(path.trim(), fsContent, name.trim(), userId);
     socket.emit("file:update:ack", { path, name, success });
   });
 
@@ -39,8 +40,10 @@ export const projectSocket = (socket: Socket) => {
   // Remove a file
   socket.on("file:remove", async (req) => {
     const { path, name } = req;
+    const userId = socket.data.userId;
     try {
-      const result = await ProjectFile.deleteOne({ path, name });
+      const result = await ProjectFile.deleteOne({ path, name, userId });
+      console.log(path,name,result)
       socket.emit("file:remove:ack", { path, name, success: result.deletedCount > 0 });
     } catch (err) {
       console.error(err);
@@ -51,11 +54,11 @@ export const projectSocket = (socket: Socket) => {
   // Rename a file
   socket.on("file:rename", async (req) => {
     const { path, oldName, name } = req;
+    const userId = socket.data.userId;
     try {
       const result = await ProjectFile.findOneAndUpdate(
-        { path, name: oldName },
+        { path, name: oldName, userId },
         { name },
-        { new: true }
       );
       socket.emit("file:rename:ack", { path, oldName, name, success: !!result });
     } catch (err) {
